Moved populate_maplayers, adjusted import statement in tests.py.
